refactor(newsletter): remove tsx, simplify html output
